Review fixes

+++ src/cluster.c
@@ -771,7 +771,7 @@ static void MR_RefreshClusterData(){
         // We need to get the port using the `RedisModule_GetClusterNodeInfo` API because on 7.2
         // invoking `cluster slot` from RM_Call will always return the none tls port.
         // For for information refer to: https://github.com/redis/redis/pull/12233
-        long long port = 0;
+        int port = 0;
         RedisModule_ThreadSafeContextLock(mr_staticCtx);
         RedisModule_GetClusterNodeInfo(mr_staticCtx, nodeId, NULL, NULL, &port, NULL);
         RedisModule_ThreadSafeContextUnlock(mr_staticCtx);
